Rename project to infrahub_mcp instead of infrahub_mcp_server
